requirements.txt
fastapi
sse-starlette
requests
beautifulsoup4
langchain-community
langchain-text-splitters
langchain-chroma
langchain-huggingface
langchain-openai
openai
pymongo
faiss-cpu
chromadb
motor
uuid